+++ packages/ai-constructs/src/conversation/conversation_handler_construct.ts
@@ -117,7 +117,10 @@ export class ConversationHandlerFunction
       conversationHandler.addToRolePolicy(
         new PolicyStatement({
           effect: Effect.ALLOW,
-          actions: ['bedrock:InvokeModel'],
+          actions: [
+            'bedrock:InvokeModel',
+            'bedrock:InvokeModelWithResponseStream',
+          ],
           resources,
         })
       );