+++ packages/deployed-backend-client/src/deployed_backend_client_list_sandboxes.test.ts
@@ -7,14 +7,8 @@ import {
   StackStatus,
 } from '@aws-sdk/client-cloudformation';
 import { BackendDeploymentStatus } from './deployed_backend_client_factory.js';
-import { platformOutputKey } from '@aws-amplify/backend-output-schemas';
 import { DefaultBackendOutputClient } from './backend_output_client.js';
 import { DefaultDeployedBackendClient } from './deployed_backend_client.js';
-import {
-  BackendOutputClientError,
-  BackendOutputClientErrorType,
-  StackIdentifier,
-} from './index.js';
 import { AmplifyClient } from '@aws-sdk/client-amplify';
 import { S3 } from '@aws-sdk/client-s3';
 import { DeployedResourcesEnumerator } from './deployed-backend-client/deployed_resources_enumerator.js';
@@ -34,14 +28,6 @@ const listStacksMock = {
   ],
 };
 
-const getOutputMockResponse = {
-  [platformOutputKey]: {
-    payload: {
-      deploymentType: 'sandbox',
-    },
-  },
-};
-
 void describe('Deployed Backend Client list sandboxes', () => {
   const mockCfnClient = new CloudFormation();
   const mockS3Client = new S3();
@@ -56,9 +42,18 @@ void describe('Deployed Backend Client list sandboxes', () => {
         const matchingStack = listStacksMock.StackSummaries.find((stack) => {
           return stack.StackName === request.input.StackName;
         });
-        const stack = matchingStack;
         return {
-          Stacks: [stack],
+          Stacks: [
+            {
+              ...matchingStack,
+              Tags: [
+                {
+                  Key: 'amplify:deployment-type',
+                  Value: 'sandbox',
+                },
+              ],
+            },
+          ],
         };
       }
       throw request;
@@ -84,23 +79,7 @@ void describe('Deployed Backend Client list sandboxes', () => {
     mockCfnClient,
     new AmplifyClient()
   );
-  const getOutputMock = mock.method(
-    mockBackendOutputClient,
-    'getOutput',
-    (backendIdentifier: StackIdentifier) => {
-      if (backendIdentifier.stackName === 'amplify-test-not-a-sandbox') {
-        return {
-          ...getOutputMockResponse,
-          [platformOutputKey]: {
-            payload: {
-              deploymentType: 'branch',
-            },
-          },
-        };
-      }
-      return getOutputMockResponse;
-    }
-  );
+
   const returnedSandboxes = [
     {
       deploymentType: 'sandbox',
@@ -117,7 +96,6 @@ void describe('Deployed Backend Client list sandboxes', () => {
   ];
 
   beforeEach(() => {
-    getOutputMock.mock.resetCalls();
     listStacksMockFn.mock.resetCalls();
     cfnClientSendMock.mock.resetCalls();
     const deployedResourcesEnumerator = new DeployedResourcesEnumerator(
@@ -209,57 +187,36 @@ void describe('Deployed Backend Client list sandboxes', () => {
     assert.equal(listStacksMockFn.mock.callCount(), 2);
   });
 
-  void it('paginates listBackends when one page contains a stack, but it gets filtered due to not having gen2 outputs', async () => {
-    getOutputMock.mock.mockImplementationOnce(() => {
-      throw new BackendOutputClientError(
-        BackendOutputClientErrorType.METADATA_RETRIEVAL_ERROR,
-        'Test metadata retrieval error'
-      );
-    });
-    listStacksMockFn.mock.mockImplementationOnce(() => {
-      return {
-        StackSummaries: [
-          {
-            StackName: 'amplify-test-name-sandbox-testHash',
-            StackStatus: StackStatus.CREATE_COMPLETE,
-            CreationTime: new Date(0),
-            LastUpdatedTime: new Date(1),
-          },
-        ],
-        NextToken: 'abc',
-      };
-    });
+  void it('filter stacks that do not have deploymentType tag in it', async () => {
+    cfnClientSendMock.mock.mockImplementation(
+      (request: ListStacksCommand | DescribeStacksCommand) => {
+        if (request instanceof ListStacksCommand) {
+          return listStacksMockFn(request.input);
+        }
+        if (request instanceof DescribeStacksCommand) {
+          const matchingStack = listStacksMock.StackSummaries.find((stack) => {
+            return stack.StackName === request.input.StackName;
+          });
+          return {
+            Stacks: [
+              {
+                ...matchingStack,
+                Tags: [],
+              },
+            ],
+          };
+        }
+        throw request;
+      }
+    );
     const sandboxes = deployedBackendClient.listBackends({
       deploymentType: 'sandbox',
     });
     assert.deepEqual(
-      (await sandboxes.getBackendSummaryByPage().next()).value,
-      returnedSandboxes
+      (await sandboxes.getBackendSummaryByPage().next()).done,
+      true
     );
 
-    assert.equal(listStacksMockFn.mock.callCount(), 2);
-  });
-
-  void it('does not paginate listBackends when one page throws an unexpected error fetching gen2 outputs', async () => {
-    getOutputMock.mock.mockImplementationOnce(() => {
-      throw new Error('Unexpected Error!');
-    });
-    listStacksMockFn.mock.mockImplementationOnce(() => {
-      return {
-        StackSummaries: [
-          {
-            StackName: 'amplify-test-name-sandbox-testHash',
-            StackStatus: StackStatus.CREATE_COMPLETE,
-            CreationTime: new Date(0),
-            LastUpdatedTime: new Date(1),
-          },
-        ],
-        NextToken: 'abc',
-      };
-    });
-    const listBackendsPromise = deployedBackendClient.listBackends({
-      deploymentType: 'sandbox',
-    });
-    await assert.rejects(listBackendsPromise.getBackendSummaryByPage().next());
+    assert.equal(listStacksMockFn.mock.callCount(), 1);
   });
 });