+++ packages/backend-ai/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @aws-amplify/backend-ai
 
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies [b6761b0]
+  - @aws-amplify/ai-constructs@0.6.0
+
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [46a0e85]
+- Updated dependencies [faacd1b]
+  - @aws-amplify/ai-constructs@0.5.0
+
 ## 0.3.0
 
 ### Minor Changes